usb/usbcdcethernet: Use asynchronous USB requests
On the bulk endpoints before synchronous receive requests were used,
which led to blocking, if no received frames were available on
the Raspberry Pi 4 and 5. This did not work. Now ReceiveFrame() returns
quickly in any case, so the network stack can continue to work, even
when there are no frames received (untested).