usb/dwhcidevice: Add vendor ID, provided by QEMU
QEMU is using a vendor ID for the USB host controller emulation for
Raspberry Pi 1-3, which is different from the ID of the real hardware.
Add this ID to the allowed IDs, so that Circle USB can be used with
unpatched QEMU versions. A patch is still necessary for the USB CDC
Ethernet emulation and maybe for USB bulk endpoint support in other
drivers.